requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.1
langchain>=0.3.0
langchain-community>=0.3.0
chromadb>=0.4.18
pypdf2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
python-pptx>=0.6.21
striprtf>=0.0.26
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.6.0
huggingface-hub>=0.19.4
requests>=2.31.0
sqlalchemy>=2.0.0
alembic>=1.12.0
tabulate>=0.9.0
langchain-text-splitters>=0.2.0
nltk>=3.8.1
spacy>=3.7.0

